+++ src/main/java/org/ldbcouncil/finbench/driver/workloads/transaction/queries/ComplexRead9.java
@@ -31,14 +31,14 @@ public class ComplexRead9 extends LdbcOperation<List<ComplexRead9Result>> {
     public static final String TRUNCATION_ORDER = "truncationOrder";
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
     private final long id;
-    private final long threshold;
+    private final double threshold;
     private final Date startTime;
     private final Date endTime;
     private final int truncationLimit;
     private final TruncationOrder truncationOrder;
 
     public ComplexRead9(@JsonProperty(ID) long id,
-                        @JsonProperty(THRESHOLD) long threshold,
+                        @JsonProperty(THRESHOLD) double threshold,
                         @JsonProperty(START_TIME) Date startTime,
                         @JsonProperty(END_TIME) Date endTime,
                         @JsonProperty(TRUNCATION_LIMIT) int truncationLimit,
@@ -69,7 +69,7 @@ public class ComplexRead9 extends LdbcOperation<List<ComplexRead9Result>> {
         return id;
     }
 
-    public long getThreshold() {
+    public double getThreshold() {
         return threshold;
     }
